% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overviewGarminRunning.R
\name{overviewGarminRunning}
\alias{overviewGarminRunning}
\title{overviewGarminRunning}
\usage{
overviewGarminRunning(
  data,
  date_from = NA,
  plot,
  target_time = median(data$Avg.Pace)
)
}
\arguments{
\item{date_from}{a date processed by as.Date e.g as.Date("2020-03-18"). This is a cut-off data for your analysis. Everything after this will be included in the analysis.}

\item{plot}{can be one of "runs_per_loc_per_dist", "distance_per_location", "mean_pace", "pca_plot", and "total_runs"}

\item{target_time}{a pace in the form "7:30" (character variable) with "minutes:seconds" that is one previously achieved that you wish to mark on your plots, OR it could be a target pace you have set yourself}

\item{output}{from the processGarminRunning function}
}
\value{
plot
}
\description{
A function to generate stylish plots that rapidly give an overview of your running.
}
\examples{
#What about my pace - calculate my mean avg.pace amongst all my runs per location

overviewGarminRunning(my_runs,
                     date_from = as.Date("2020-03-18"),
                     plot = "mean_pace",
                     target_time = "7:18")

#Note the 5-10 mile runs the pace is slightly faster in manchester than in Bromsgrove probably because of the hills!

# we can also utilise numeric data to do PCA analysis - what are the major sources of variation in our running data?
overviewGarminRunning(my_runs,
                     date_from = as.Date("2020-03-18"),
                     plot = "pca_plot",
                     target_time = "7:18")

}
